+++ log/src/robocorp/log/__init__.py
@@ -35,7 +35,7 @@ from .protocols import IReadLines, LogHTMLStyle, Status
 if typing.TYPE_CHECKING:
     from ._robo_logger import _RoboLogger
 
-__version__ = "2.5.0"
+__version__ = "2.6.0"
 version_info = [int(x) for x in __version__.split(".")]